refactor(deploy): :recycle: move helpers to utils
