+++ scripts/1_deploy_contracts.js
@@ -1,17 +1,12 @@
 require('dotenv').config()
 const { ethers, upgrades } = require('hardhat')
-
-// Helper function to print titles in a box format
-function printBoxedTitle (title) {
-  const line = '='.repeat(title.length + 4)
-  console.log(`\n+${line}+\n|  ${title}  |\n+${line}+\n`)
-}
-
-// Fetch the admin address
-async function getAdminAddress () {
-  const [deployer] = await ethers.getSigners()
-  return process.env.ADMIN_ADDRESS || deployer.address
-}
+const {
+  getAdminAddress,
+  printBoxedTitle,
+  verifyImplementationContract,
+  verifyProxyAdminContract,
+  verifyProxyContract
+} = require('./utils')
 
 // Initialize contract arguments
 function getInitializerArguments (admin) {
@@ -43,46 +38,15 @@ async function deployCMTATBase () {
   return proxyContract
 }
 
-// Verify a contract on Etherscan
-async function verifyOnExplorer (contractAddress, constructorArguments = []) {
-  console.log('Verifying contract on explorer...')
-  await run('verify:verify', { address: contractAddress, constructorArguments })
-  console.log(`Verification submitted for contract at address: ${contractAddress}`)
-}
-
-// Verify different components of the contract
-async function verifyProxyContract (proxyContract) {
-  printBoxedTitle('Verifying Proxy Contract')
-  const proxyAddress = await proxyContract.getAddress()
-  console.log('Proxy contract address:', proxyAddress, '\n')
-  await verifyOnExplorer(proxyAddress, getInitializerArguments(await getAdminAddress()))
-
-  // Note: Verification of the proxy contract may be unnecessary but is included for completeness.
-}
-
-async function verifyProxyAdminContract (proxyContract) {
-  printBoxedTitle('Verifying Proxy Admin Contract')
-  const proxyAdminAddress = await upgrades.erc1967.getAdminAddress(await proxyContract.getAddress())
-  console.log('Proxy Admin contract address:', proxyAdminAddress, '\n')
-  await verifyOnExplorer(proxyAdminAddress)
-
-  // Note: Verification of the proxy admin contract may be unnecessary but is included for completeness.
-}
-
-async function verifyImplementationContract (proxyContract) {
-  printBoxedTitle('Verifying Implementation Contract')
-  const implementationAddress = await upgrades.erc1967.getImplementationAddress(await proxyContract.getAddress())
-  console.log('Implementation contract address:', implementationAddress, '\n')
-  await verifyOnExplorer(implementationAddress)
-}
-
 // Main verification function
 async function verifyContract (proxyContract) {
   const delay = 30000 // 30-second delay for network propagation
   console.log(`\nWaiting ${delay / 1000} seconds for the network to propagate...`)
   await new Promise(resolve => setTimeout(resolve, delay))
 
-  await verifyProxyContract(proxyContract)
+  const constructorArguments = getInitializerArguments(await getAdminAddress())
+
+  await verifyProxyContract(proxyContract, constructorArguments)
   await verifyProxyAdminContract(proxyContract)
   await verifyImplementationContract(proxyContract)
 }
